refactor(engine): reorganize folder structure into logical modules
Move flat engine files into core/, stochastic/, scheduler/, nodes/, and
validation/ subdirectories with co-located __tests__/ folders. Update all
import paths accordingly.
